+++ semana7/aula27-integração-API/cadastro/src/pages/Register.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
+import axios from 'axios'
 
-const Container = styled.div `
+const Container = styled.div`
     padding: 4rem 2rem 1rem 2rem;
     border-radius: 10px;
     box-shadow: 0 0 10px #333;
@@ -12,8 +13,7 @@ const Container = styled.div `
     align-items: center;
 `
 
-
-const Inputs= styled.input `
+const Inputs = styled.input`
     padding: 0.5rem;
     margin: 0.5rem;
     font-size: 1rem;
@@ -28,11 +28,10 @@ const Inputs= styled.input `
         box-shadow: 0 0 0 0;
     }
 `
-
 const InputName = styled(Inputs)``
 const InputEmail = styled(Inputs)``
 
-const Button = styled.button `
+const Button = styled.button`
     padding: 0.4rem 1rem;
     margin: 1rem;
     font-size: 1rem;
@@ -50,12 +49,12 @@ const Button = styled.button `
     }
 `
 
-const ButtonAdd = styled(Button) `
+const ButtonAdd = styled(Button)`
     background: #733236;
     color: #E3DBD3;
 `
 
-const ButtonScreen = styled(Button) `
+const ButtonScreen = styled(Button)`
     margin-top: 2rem;
     background: #E3DBD3;
     color: #733236;
@@ -63,6 +62,45 @@ const ButtonScreen = styled(Button) `
 `
 
 export default class Register extends React.Component {
+    state = {
+        inputName: '',
+        inputEmail: ''
+    }
+
+    onChangeName = (e) => {
+        this.setState({ inputName: e.target.value })
+    }
+    onChangeEmail = (e) => {
+        this.setState({ inputEmail: e.target.value })
+    }
+
+    createUser = async () => {
+        const body = {
+            name: this.state.inputName,
+            email: this.state.inputEmail
+        }
+        try {
+            const response = await axios.post('https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users', body, {
+                headers: {
+                    Authorization: 'adryane-fernandes-cruz'
+                }
+            })
+
+            this.getUsers()
+            this.setState({ inputEmail: '', inputName: '' })
+            this.setState({ page: 'list' })
+            alert('Usuário criado :)')
+            
+        } catch (err) {
+            if (!this.state.inputName || !this.state.inputEmail) {
+                alert('Por favor, preencha todos os campos corretamente.')
+            } else {
+                console.log(err.response.data)
+                alert('Erro ao criar usuário!')
+            }
+        }
+    }
+
     render() {
         return (
             <Container>
